.vecchiume/requirements.txt
# needed to generate image gallery
imagesize
pyyaml